# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Velzo Digital Marketplace API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  buyer: "buyer",
  seller: "seller",
} as const;

export interface RegisterBody {
  email: string;
  password: string;
  name: string;
  role: RegisterBodyRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  role: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  bio?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Store {
  id: number;
  userId: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  logo?: string | null;
  /** @nullable */
  banner?: string | null;
  owner?: User;
  productCount: number;
  createdAt: string;
}

export interface StoreOrNull {
  store: Store | null;
}

export interface CreateStoreBody {
  name: string;
  description?: string;
  logo?: string;
  banner?: string;
}

export interface UpdateStoreBody {
  name?: string;
  description?: string;
  logo?: string;
  banner?: string;
}

export interface Product {
  id: number;
  storeId: number;
  /** @nullable */
  storeName?: string | null;
  title: string;
  description: string;
  price: number;
  category: string;
  /** @nullable */
  coverImage?: string | null;
  /** @nullable */
  fileUrl?: string | null;
  status: string;
  reviewCount: number;
  avgRating: number;
  createdAt: string;
}

export interface CreateProductBody {
  title: string;
  description: string;
  price: number;
  category: string;
  coverImage?: string;
  fileUrl?: string;
}

export interface UpdateProductBody {
  title?: string;
  description?: string;
  price?: number;
  category?: string;
  coverImage?: string;
  fileUrl?: string;
  status?: string;
}

export interface Review {
  id: number;
  productId: number;
  userId: number;
  userName: string;
  /** @nullable */
  userAvatar?: string | null;
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface CreateReviewBody {
  rating: number;
  comment?: string;
}

export interface Conversation {
  id: number;
  otherUser: User;
  /** @nullable */
  lastMessage?: string | null;
  /** @nullable */
  lastMessageAt?: string | null;
  unreadCount: number;
  createdAt: string;
}

export interface CreateConversationBody {
  recipientId: number;
}

export interface Message {
  id: number;
  conversationId: number;
  senderId: number;
  senderName: string;
  /** @nullable */
  senderAvatar?: string | null;
  content: string;
  createdAt: string;
}

export interface SendMessageBody {
  content: string;
}

export interface CatalogStats {
  totalProducts: number;
  totalStores: number;
  totalCategories: number;
  totalSellers: number;
}

export type ListProductsParams = {
  category?: string;
  search?: string;
  sort?: string;
  minPrice?: number;
  maxPrice?: number;
};